+++ bin/template/lib/behaviors.dart
@@ -1,3 +1,2 @@
 library __projectName__.behaviors;
 
-export './src/behaviors/test_behavior.dart';