+++ bin/template/lib/models.dart
@@ -1,3 +1,2 @@
 library __projectName__.models;
 
-export 'src/models/my_model.dart';